refactor(bgrec): split nested state classes into BackgroundRecorder.State.cs
